Use better suited integer type for keyboard keys

+++ source/swift/Keyboard/Keyboard.Hotkey.swift
@@ -2,15 +2,15 @@ import Foundation
 
 public struct KeyboardHotkey: Equatable
 {
-    public var key: UInt32
+    public var key: UInt16
     public var modifier: UInt32
 
-    public init(key: UInt32, modifier: UInt32) {
+    public init(key: UInt16, modifier: UInt32) {
         self.key = key
         self.modifier = modifier
     }
 
-    public init(key: UInt32, modifier: KeyboardModifier) {
+    public init(key: UInt16, modifier: KeyboardModifier) {
         self.key = key
         self.modifier = modifier.rawValue
     }
